Fixed: Job runTime not updated in jobConfiguration component on skip action(#2a7yg9j)
